+++ core/src/main/kotlin/cache/data/GuildOnboardingData.kt
@@ -1,17 +1,8 @@
 package dev.kord.core.cache.data
 
-import dev.kord.common.entity.DiscordGuildOnboarding
-import dev.kord.common.entity.Snowflake
+import dev.kord.common.entity.*
 import kotlinx.serialization.Serializable
 
-/**
- * Represents a Guild Onboarding object
- *
- * @param guildId The ID of the guild this onboarding is part of
- * @param prompts Prompts show during onboarding and in customize community
- * @param defaultChannelIds Channel IDs that members get opted into automatically
- * @param enabled Whether onboarding is enabled in the guild
- */
 @Serializable
 public data class GuildOnboardingData(
     val guildId: Snowflake,
@@ -26,6 +17,54 @@ public data class GuildOnboardingData(
                 prompts.map { OnboardingPromptData.from(it) },
                 defaultChannelIds,
                 enabled
+            )
+        }
+    }
+}
+
+@Serializable
+public data class OnboardingPromptData(
+        val id: Snowflake,
+        val type: OnboardingPromptType,
+        val options: List<OnboardingPromptOptionData>,
+        val title: String,
+        val singleSelect: Boolean,
+        val required: Boolean,
+        val inOnboarding: Boolean
+) {
+    public companion object {
+        public fun from(entity: DiscordOnboardingPrompt): OnboardingPromptData = with(entity) {
+            OnboardingPromptData(
+                    id,
+                    type,
+                    options.map { OnboardingPromptOptionData.from(it) },
+                    title,
+                    singleSelect,
+                    required,
+                    inOnboarding
+            )
+        }
+    }
+}
+
+@Serializable
+public data class OnboardingPromptOptionData(
+        val id: Snowflake,
+        val channelIds: List<Snowflake>,
+        val roleIds: List<Snowflake>,
+        val emoji: DiscordEmoji,
+        val title: String,
+        val description: String?
+) {
+    public companion object {
+        public fun from(entity: DiscordOnboardingPromptOption): OnboardingPromptOptionData = with(entity) {
+            OnboardingPromptOptionData(
+                    id,
+                    channelIds,
+                    roleIds,
+                    emoji,
+                    title,
+                    description
             )
         }
     }
